requirements.txt
django
djangorestframework
google-generativeai
gspread
oauth2client
python-dotenv
python-dateutil
gunicorn
requests
